refactor(core): move org resolver to utils/ (was inappropriately in core/)
